Cleanup: * move general purpose mex conversion functions into their own header * add documentation to mexify
