+++ drake/util/mexify.h
@@ -11,6 +11,37 @@
 #include "mex.h"
 #include "drakeMexUtil.h"
 
+/**
+ * Functions that make it easy to create a mex function given a C++ function signature in the form of a std::function.
+ *
+ * Example:
+ * Calling a C++ function with signature
+ *
+ *   bool foo(int arg)
+ *
+ * can be done by creating a mex function with body
+ *
+ *   function<bool(int)> func = &foo;
+ *   mexCallFunction(func, nlhs, plhs, nrhs, prhs);
+ *
+ * Overloads for the fromMex (resp. toMex) function should be provided for each argument type (resp. return type).
+ *
+ * fromMex functions should be of the form
+ *
+ *   ArgumentType fromMex(const mxArray* source, TriggerType*)
+ *
+ * where TriggerType is the type of the argument currently being processed according to the signature
+ * of the std::function that was passed in (with cv qualifiers and reference removed)
+ * and ArgumentType is the type that will actually be passed into the function, which may or may not be the same as TriggerType.
+ * romMex functions should throw a MexToCppConversionError if the 'source' argument cannot be processed
+ *
+ * toMex functions should be of the form
+ *
+ *   void toMex(const SourceType& source, mxArray* dest[], int nlhs)
+ *
+ * where SourceType is such that the return type of the std::function can be converted to it.
+ */
+
 class MexToCppConversionError : public std::runtime_error {
 public:
   MexToCppConversionError(const std::string& msg) : std::runtime_error(msg) { };
@@ -18,29 +49,6 @@ public:
   MexToCppConversionError(const char *msg) : std::runtime_error(msg) { };
 };
 
-/**
- * fromMex functions should be of the form
- * ReturnType fromMex(const mxArray* source, ArgumentType*)
- *
- * toMex functions should be of the form
- * void toMex(const SourceType& source, mxArray* dest[], int nlhs)
- */
-
-int fromMex(const mxArray* source, int*) {
-  if (mxGetM(source) != 1 || mxGetN(source) != 1)
-    throw MexToCppConversionError("Expected scalar.");
-  return static_cast<int>(mxGetScalar(source));
-}
-
-bool fromMex(const mxArray* source, bool*) {
-  if (!mxIsLogicalScalar(source))
-    throw MexToCppConversionError("Expected logical.");
-  return mxGetLogicals(source)[0];
-}
-
-/**
- * mexCallFunction
- */
 class MexCallFunctionError : public std::runtime_error {
 public:
   MexCallFunctionError(const std::string& msg) : std::runtime_error(msg) { };
@@ -48,14 +56,17 @@ public:
   MexCallFunctionError(const char *msg) : std::runtime_error(msg) { };
 };
 
-// base case
+/**
+ * mexCallFunction
+ */
+// base case with return
 template <typename R>
 void mexCallFunction(std::function<R(void)> func, int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], int nrhs_already_processed) {
   // call function and convert to mxArray
   toMex(func(), plhs, nlhs);
 }
 
-// void return type specialization
+// base case without return
 template <>
 void mexCallFunction(std::function<void(void)> func, int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], int nrhs_already_processed) {
   func();
@@ -66,7 +77,7 @@ template<typename R, typename Arg0, typename ...Args>
 void mexCallFunction(std::function<R(Arg0, Args...)> func, int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], int nrhs_already_processed = 0)
 {
   // check number of input arguments
-  const int expected_num_args = sizeof...(Args) + 1;
+  const int expected_num_args = sizeof...(Args) + 1; // don't forget Arg0
   if (nrhs != expected_num_args) {
     std::ostringstream buf;
     buf << "Expected " << expected_num_args << " arguments, but got " << nrhs << ".";
@@ -124,11 +135,14 @@ public:
   }
 };
 
+// recursive case
 template <typename Func0, typename ...Funcs>
 void mexTryToCallFunctions(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], Func0 function0, Funcs... functions)
 {
   try {
     mexCallFunction(function0, nlhs, plhs, nrhs, prhs);
+    // Debug:
+    // std::cout << "mexTryToCallFunctions called function:\n" << typeid(function0).name() << std::endl;
   }
   catch (MexCallFunctionError& call_function_error) {
     try {
@@ -141,12 +155,12 @@ void mexTryToCallFunctions(int nlhs, mxArray *plhs[], int nrhs, const mxArray *p
   }
 }
 
+// base case: no more functions left to try
 template <typename Func0>
 void mexTryToCallFunctions(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[], Func0 function0)
 {
   try {
     mexCallFunction(function0, nlhs, plhs, nrhs, prhs);
-
     // Debug:
     // std::cout << "mexTryToCallFunctions called function:\n" << typeid(function0).name() << std::endl;
   }
@@ -156,6 +170,5 @@ void mexTryToCallFunctions(int nlhs, mxArray *plhs[], int nrhs, const mxArray *p
     throw options_exhausted_error;
   }
 }
-
 
 #endif //DRAKE_MEXIFY_H
